+++ lib/dune/opts.ex
@@ -42,6 +42,10 @@ defmodule Dune.Opts do
     Use pretty printing when inspecting the result.
     Should be a boolean. Defaults to `false`.
 
+  - `inspect_sort_maps`:
+    Sort maps when inspecting the result, useful to keep the output deterministic.
+    Should be a boolean. Defaults to `false`.
+
   ### Extra note about `atom_pool_size`
 
   Atoms are reused from one evaluation to the other so the total is not
@@ -66,7 +70,8 @@ defmodule Dune.Opts do
           max_heap_size: pos_integer,
           max_reductions: pos_integer,
           timeout: pos_integer,
-          pretty: boolean
+          pretty: boolean,
+          inspect_sort_maps: boolean
         }
 
   defstruct atom_pool_size: 5000,
@@ -75,7 +80,8 @@ defmodule Dune.Opts do
             max_heap_size: 50_000,
             max_reductions: 30_000,
             timeout: 50,
-            pretty: false
+            pretty: false,
+            inspect_sort_maps: false
 
   @doc """
   Validates untrusted options from a keyword or a map and returns a `Dune.Opts` struct.
@@ -159,6 +165,10 @@ defmodule Dune.Opts do
 
   defp do_validate(%{pretty: pretty}) when not is_boolean(pretty) do
     raise ArgumentError, message: "pretty should be a boolean"
+  end
+
+  defp do_validate(%{inspect_sort_maps: sort}) when not is_boolean(sort) do
+    raise ArgumentError, message: "inspect_sort_maps should be a boolean"
   end
 
   defp do_validate(opts = %{allowlist: allowlist}) do