+++ lib/dune/parser/unsafe_ast.ex
@@ -1,9 +1,8 @@
 defmodule Dune.Parser.UnsafeAst do
   @moduledoc false
 
-  @type atom_mapping :: {atom, String.t()}
   @type t :: %__MODULE__{
-          ast: String.t(),
+          ast: Macro.t(),
           atom_mapping: Dune.AtomMapping.t()
         }
   @enforce_keys [:ast, :atom_mapping]